refactor(test):테스트 초기 설정을 상위 클래스에서 하도록 수정
