+++ src/test/java/wooteco/subway/section/SectionCreateAcceptanceTest.java
@@ -19,50 +19,11 @@ import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-
 public class SectionCreateAcceptanceTest extends AcceptanceTest {
 
     private final StationResponse gangnam = new StationResponse(1L, "강남역");
     private final StationResponse yeoksam = new StationResponse(2L, "역삼역");
     private final StationResponse seolleung = new StationResponse(3L, "선릉역");
-
-    @BeforeEach
-    @Override
-    public void setUp() {
-        super.setUp();
-
-        Map<String, String> stationParams = new HashMap<>();
-        stationParams.put("name", "강남역");
-        RestAssuredHelper.jsonPost(stationParams, "/stations");
-
-        Map<String, String> stationParams2 = new HashMap<>();
-        stationParams2.put("name", "역삼역");
-        RestAssuredHelper.jsonPost(stationParams2, "/stations");
-
-        Map<String, String> lineParams = new HashMap<>();
-        lineParams.put("color", "bg-red-600");
-        lineParams.put("name", "신분당선");
-        lineParams.put("upStationId", "1");
-        lineParams.put("downStationId", "2");
-        lineParams.put("distance", "10");
-        RestAssuredHelper.jsonPost(lineParams, "/lines");
-
-        Map<String, String> stationParams3 = new HashMap<>();
-        stationParams3.put("name", "선릉역");
-        RestAssuredHelper.jsonPost(stationParams3, "/stations");
-
-        Map<String, String> stationParams4 = new HashMap<>();
-        stationParams4.put("name", "삼성역");
-        RestAssuredHelper.jsonPost(stationParams4, "/stations");
-
-        Map<String, String> lineParams2 = new HashMap<>();
-        lineParams2.put("color", "bg-green-600");
-        lineParams2.put("name", "2호선");
-        lineParams2.put("upStationId", "3");
-        lineParams2.put("downStationId", "4");
-        lineParams2.put("distance", "10");
-        RestAssuredHelper.jsonPost(lineParams2, "/lines");
-    }
 
     @DisplayName("구간 생성 성공 - 상행 종점역")
     @Test
