+++ src/test/java/wooteco/subway/station/StationAcceptanceTest.java
@@ -1,7 +1,9 @@
 package wooteco.subway.station;
 
+import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
@@ -19,6 +21,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
+
+    @BeforeEach
+    @Override
+    public void setUp() {
+        RestAssured.port = super.port;
+    }
 
     @DisplayName("지하철역 생성 성공")
     @Test
